requirements.txt
fastapi
uvicorn
python-multipart
facenet-pytorch
opencv-python
torch
torchvision
numpy
Pillow
supabase
python-dotenv
pydantic
aiofiles
jinja2